solution/requirements.txt
django==2.2.5
djangorestframework==3.10.3
requests==2.22.0
flake8==3.7.8
markdown==3.1.1